backend/pyproject.toml
[project]
name = "smart-error-book-backend"
version = "0.1.0"
description = "Smart Error Book - Backend API Service"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI 核心
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # 数据库
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",  # PostgreSQL 异步驱动
    "psycopg2-binary>=2.9.9",
    
    # Redis
    "redis>=5.0.1",
    "hiredis>=2.3.2",
    
    # 认证和安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    
    # 数据验证
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # AI 相关
    "openai>=1.10.0",
    "anthropic>=0.8.0",
    "tiktoken>=0.5.2",
    "sentence-transformers>=2.3.1",
    
    # 工具库
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "python-slugify>=8.0.1",
    "pillow>=10.2.0",
    
    # 日志和监控
    "loguru>=0.7.2",
    "prometheus-client>=0.19.0",
    
    # 任务队列
    "celery>=5.3.4",
    "arq>=0.25.0",
    
    # OCR
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

